renamed bookstores' files such that: Book and other related files are prefixed with BS also eg: Book -> BSBook
